유저 페이징 처리 완료

+++ src/main/java/com/nexters/wiw/api/service/UserService.java
@@ -1,17 +1,16 @@
 package com.nexters.wiw.api.service;
 
-import java.util.List;
-
 import com.nexters.wiw.api.domain.User;
 import com.nexters.wiw.api.domain.UserRepository;
 import com.nexters.wiw.api.domain.error.ErrorType;
-import com.nexters.wiw.api.exception.BadRequestException;
 import com.nexters.wiw.api.exception.NotFoundException;
 import com.nexters.wiw.api.exception.UnAuthorizedException;
 import com.nexters.wiw.api.ui.UserPatchRequestDto;
 import com.nexters.wiw.api.ui.UserRequestDto;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -28,7 +27,7 @@ public class UserService {
     @Autowired
     private PasswordEncoder bCryptPasswordEncoder;
 
-    public User getOne(String authHeader, Long id) {
+    public User getOne(final String authHeader, Long id) {
         if (!authService.isValidateToken(authHeader))
             throw new UnAuthorizedException(ErrorType.UNAUTHORIZED, "유효하지 않은 토큰입니다.");
 
@@ -36,15 +35,16 @@ public class UserService {
                 .orElseThrow(() -> new NotFoundException(ErrorType.NOT_FOUND, "아이디에 해당하는 유저가 존재하지 않습니다."));
     }
 
-    public List<User> getList(String authHeader, String query) {
+    public Page<User> getList(final String authHeader, final Pageable pageable, final String query) {
         if (!authService.isValidateToken(authHeader))
             throw new UnAuthorizedException(ErrorType.UNAUTHORIZED, "유효하지 않은 토큰입니다.");
 
-        // TODO 모든 유저는 과부화가 심하니 예외처리로 변경하는 건 어떤지..
+        // TODO 모든 유저는 과부화가 심하니 예외처리로 변경하는 건 어떤지?
         if(query == null) 
-            return userRepository.findAll();
+            return userRepository.findAll(pageable);
 
-        return userRepository.findByEmailContainingOrNicknameContaining(query, query);
+        // TODO @뒤의 이메일 도메인까지 검색이 되는 것 막기
+        return userRepository.findByEmailContainingOrNicknameContaining(query, query, pageable);
     }
 
     @Transactional
@@ -54,7 +54,7 @@ public class UserService {
     }
 
     @Transactional
-    public User patch(String authHeader, final Long id, final UserPatchRequestDto userPatchRequestDto) {
+    public User patch(final String authHeader, final Long id, final UserPatchRequestDto userPatchRequestDto) {
         if (!authService.isValidateToken(authHeader))
             throw new UnAuthorizedException(ErrorType.UNAUTHORIZED, "유효하지 않은 토큰입니다.");
 
@@ -62,9 +62,12 @@ public class UserService {
     }
 
     @Transactional
-    public void delete(String authHeader, final Long id) {
+    public void delete(final String authHeader, final Long id) {
         if (!authService.isValidateToken(authHeader))
             throw new UnAuthorizedException(ErrorType.UNAUTHORIZED, "유효하지 않은 토큰입니다.");
+
+        if(authService.findIdByToken(authHeader) == id)
+            throw new UnAuthorizedException(ErrorType.UNAUTHORIZED, "계정 삭제는 본인만 가능합니다.");
 
         userRepository.deleteById(id);
     }
